// gdr-web-version-mdp/gdr-web-version-mdp/src/app/@data/model/entidadGestor.ts
export class EntidadGestor {
  rolEntidadId?: number;
  rectorId: number;
  entidadId: number;
  descripcionEntidad: string;
  siglaEntidad: string;
  siglaNombreEntidad: string;
  flagAsociado: string;
  sectorId: number;
  sector: string;
  nivelGobiernoId: number;
  nivelGobierno: string;
  nro: string;
  flagCambio: string;
}